+++ main.tf
@@ -3,10 +3,14 @@ provider "azurerm" {
 }
 
 locals {
-  config_sp_image = lookup(local.config_sp_image_list, split("-", var.sharepoint_version)[0])
-  config_sp_dsc   = split("-", var.sharepoint_version)[0] == "Subscription" ? local.config_sp_se_dsc : local.config_sp_legacy_dsc
-  config_fe_dsc   = split("-", var.sharepoint_version)[0] == "Subscription" ? local.config_fe_se_dsc : local.config_fe_legacy_dsc
-  create_rdp_rule = lower(var.rdp_traffic_allowed) == "no" ? 0 : 1
+  config_sp_image                       = lookup(local.config_sp_image_list, split("-", var.sharepoint_version)[0])
+  config_sp_dsc                         = split("-", var.sharepoint_version)[0] == "Subscription" ? local.config_sp_se_dsc : local.config_sp_legacy_dsc
+  config_fe_dsc                         = split("-", var.sharepoint_version)[0] == "Subscription" ? local.config_fe_se_dsc : local.config_fe_legacy_dsc
+  create_rdp_rule                       = lower(var.rdp_traffic_allowed) == "no" ? 0 : 1
+  admin_password                        = var.admin_password == "" ? random_password.random_admin_password.result : var.admin_password
+  service_accounts_password             = var.service_accounts_password == "" ? random_password.random_service_accounts_password.result : var.service_accounts_password
+  enable_hybrid_benefit_server_licenses = var.enable_hybrid_benefit_server_licenses == true ? "Windows_Server" : "None"
+
   general_settings = {
     dscScriptsFolder      = "dsc"
     adfsSvcUserName       = "adfssvc"
@@ -33,26 +37,26 @@ locals {
 
   config_dc = {
     vmName             = "DC"
-    vmSize             = "Standard_B2s"
+    vmSize             = var.vm_dc_size
     vmImagePublisher   = "MicrosoftWindowsServer"
     vmImageOffer       = "WindowsServer"
     vmImageSKU         = "2022-datacenter-azure-edition-smalldisk"
-    storageAccountType = "Standard_LRS"
+    storageAccountType = var.vm_dc_storage_account_type
   }
 
   config_sql = {
     vmName             = "SQL"
-    vmSize             = "Standard_B2ms"
+    vmSize             = var.vm_sql_size
     vmImagePublisher   = "MicrosoftSQLServer"
     vmImageOffer       = "sql2019-ws2022"
     vmImageSKU         = "sqldev-gen2"
-    storageAccountType = "Standard_LRS"
+    storageAccountType = var.vm_sql_storage_account_type
   }
 
   config_sp = {
     vmName             = "SP"
-    vmSize             = "Standard_B4ms"
-    storageAccountType = "Standard_LRS"
+    vmSize             = var.vm_sp_size
+    storageAccountType = var.vm_sp_storage_account_type
   }
 
   config_sp_image_list = {
@@ -64,7 +68,7 @@ locals {
 
   config_fe = {
     vmName = "FE"
-    vmSize = "Standard_B4ms"
+    vmSize = var.vm_sp_size
   }
 
   config_dc_dsc = {
@@ -96,18 +100,31 @@ locals {
   }
 
   config_fe_legacy_dsc = {
-    fileName       = "ConfigureFELegacyVM.zip"
-    script         = "ConfigureFELegacyVM.ps1"
+    fileName       = "ConfigureFELegacy.zip"
+    script         = "ConfigureFELegacy.ps1"
     function       = "ConfigureFEVM"
     forceUpdateTag = "1.0"
   }
 
   config_fe_se_dsc = {
-    fileName       = "ConfigureFESEVM.zip"
-    script         = "ConfigureFESEVM.ps1"
+    fileName       = "ConfigureFESE.zip"
+    script         = "ConfigureFESE.ps1"
     function       = "ConfigureFEVM"
     forceUpdateTag = "1.0"
   }
+}
+
+# Service account password
+resource "random_password" "random_admin_password" {
+  length           = 8
+  special          = true
+  override_special = "!#$%&*()-_=+[]{}<>:?"
+}
+
+resource "random_password" "random_service_accounts_password" {
+  length           = 8
+  special          = true
+  override_special = "!#$%&*()-_=+[]{}<>:?"
 }
 
 # Create a resource group
@@ -310,8 +327,8 @@ resource "azurerm_windows_virtual_machine" "vm_dc" {
   network_interface_ids    = [azurerm_network_interface.nic_dc_0.id]
   size                     = local.config_dc["vmSize"]
   admin_username           = var.admin_username
-  admin_password           = var.admin_password
-  license_type             = "Windows_Server"
+  admin_password           = local.admin_password
+  license_type             = local.enable_hybrid_benefit_server_licenses
   timezone                 = var.time_zone
   enable_automatic_updates = true
   provision_vm_agent       = true
@@ -366,11 +383,11 @@ SETTINGS
     "configurationArguments": {
       "AdminCreds": {
         "UserName": "${var.admin_username}",
-        "Password": "${var.admin_password}"
+        "Password": "${local.admin_password}"
       },
       "AdfsSvcCreds": {
         "UserName": "${local.general_settings["adfsSvcUserName"]}",
-        "Password": "${var.service_accounts_password}"
+        "Password": "${local.service_accounts_password}"
       }
     }
   }
@@ -399,8 +416,8 @@ resource "azurerm_windows_virtual_machine" "vm_sql" {
   network_interface_ids    = [azurerm_network_interface.nic_sql_0.id]
   size                     = local.config_sql["vmSize"]
   admin_username           = "local-${var.admin_username}"
-  admin_password           = var.admin_password
-  license_type             = "Windows_Server"
+  admin_password           = local.admin_password
+  license_type             = local.enable_hybrid_benefit_server_licenses
   timezone                 = var.time_zone
   enable_automatic_updates = true
   provision_vm_agent       = true
@@ -455,15 +472,15 @@ SETTINGS
     "configurationArguments": {
       "DomainAdminCreds": {
         "UserName": "${var.admin_username}",
-        "Password": "${var.admin_password}"
+        "Password": "${local.admin_password}"
       },
       "SqlSvcCreds": {
         "UserName": "${local.general_settings["sqlSvcUserName"]}",
-        "Password": "${var.service_accounts_password}"
+        "Password": "${local.service_accounts_password}"
       },
       "SPSetupCreds": {
         "UserName": "${local.general_settings["spSetupUserName"]}",
-        "Password": "${var.service_accounts_password}"
+        "Password": "${local.service_accounts_password}"
       }
     }
   }
@@ -492,8 +509,8 @@ resource "azurerm_windows_virtual_machine" "vm_sp" {
   network_interface_ids    = [azurerm_network_interface.nic_sp_0.id]
   size                     = local.config_sp["vmSize"]
   admin_username           = "local-${var.admin_username}"
-  admin_password           = var.admin_password
-  license_type             = "Windows_Server"
+  admin_password           = local.admin_password
+  license_type             = local.enable_hybrid_benefit_server_licenses
   timezone                 = var.time_zone
   enable_automatic_updates = true
   provision_vm_agent       = true
@@ -522,7 +539,7 @@ resource "azurerm_virtual_machine_extension" "vm_sp_dsc" {
   auto_upgrade_minor_version = true
 
   timeouts {
-    create = "75m"
+    create = "120m"
   }
 
   settings = <<SETTINGS
@@ -553,39 +570,39 @@ SETTINGS
     "configurationArguments": {
       "DomainAdminCreds": {
         "UserName": "${var.admin_username}",
-        "Password": "${var.admin_password}"
+        "Password": "${local.admin_password}"
       },
       "SPSetupCreds": {
         "UserName": "${local.general_settings["spSetupUserName"]}",
-        "Password": "${var.service_accounts_password}"
+        "Password": "${local.service_accounts_password}"
       },
       "SPFarmCreds": {
         "UserName": "${local.general_settings["spFarmUserName"]}",
-        "Password": "${var.service_accounts_password}"
+        "Password": "${local.service_accounts_password}"
       },
       "SPSvcCreds": {
         "UserName": "${local.general_settings["spSvcUserName"]}",
-        "Password": "${var.service_accounts_password}"
+        "Password": "${local.service_accounts_password}"
       },
       "SPAppPoolCreds": {
         "UserName": "${local.general_settings["spAppPoolUserName"]}",
-        "Password": "${var.service_accounts_password}"
+        "Password": "${local.service_accounts_password}"
       },
       "SPADDirSyncCreds": {
         "UserName": "${local.general_settings["spADDirSyncUserName"]}",
-        "Password": "${var.service_accounts_password}"
+        "Password": "${local.service_accounts_password}"
       },
       "SPPassphraseCreds": {
         "UserName": "Passphrase",
-        "Password": "${var.service_accounts_password}"
+        "Password": "${local.service_accounts_password}"
       },
       "SPSuperUserCreds": {
         "UserName": "${local.general_settings["spSuperUserName"]}",
-        "Password": "${var.service_accounts_password}"
+        "Password": "${local.service_accounts_password}"
       },
       "SPSuperReaderCreds": {
         "UserName": "${local.general_settings["spSuperReaderName"]}",
-        "Password": "${var.service_accounts_password}"
+        "Password": "${local.service_accounts_password}"
       }
     }
   }
@@ -641,8 +658,8 @@ resource "azurerm_windows_virtual_machine" "vm_fe" {
   network_interface_ids    = [element(azurerm_network_interface.nic_fe_0.*.id, count.index)]
   size                     = local.config_sp["vmSize"]
   admin_username           = "local-${var.admin_username}"
-  admin_password           = var.admin_password
-  license_type             = "Windows_Server"
+  admin_password           = local.admin_password
+  license_type             = local.enable_hybrid_benefit_server_licenses
   timezone                 = var.time_zone
   enable_automatic_updates = true
   provision_vm_agent       = true
@@ -672,7 +689,7 @@ resource "azurerm_virtual_machine_extension" "vm_fe_dsc" {
   auto_upgrade_minor_version = true
 
   timeouts {
-    create = "90m"
+    create = "120m"
   }
 
   settings = <<SETTINGS
@@ -703,19 +720,19 @@ SETTINGS
     "configurationArguments": {
       "DomainAdminCreds": {
         "UserName": "${var.admin_username}",
-        "Password": "${var.admin_password}"
+        "Password": "${local.admin_password}"
       },
       "SPSetupCreds": {
         "UserName": "${local.general_settings["spSetupUserName"]}",
-        "Password": "${var.service_accounts_password}"
+        "Password": "${local.service_accounts_password}"
       },
       "SPFarmCreds": {
         "UserName": "${local.general_settings["spFarmUserName"]}",
-        "Password": "${var.service_accounts_password}"
+        "Password": "${local.service_accounts_password}"
       },
       "SPPassphraseCreds": {
         "UserName": "Passphrase",
-        "Password": "${var.service_accounts_password}"
+        "Password": "${local.service_accounts_password}"
       }
     }
   }